mall：完善收件地址

+++ yudao-module-member/yudao-module-member-biz/src/main/java/cn/iocoder/yudao/module/member/dal/dataobject/address/AddressDO.java
@@ -39,7 +39,7 @@ public class AddressDO extends BaseDO {
     /**
      * 地区编号
      */
-    private Integer areaId;
+    private Long areaId;
     /**
      * 邮编
      */
